Add -DINTEL_CUSTOMIZATION=1 option for CMAKE_C_CFLAGS (#8501)
The compiler-rt files are a mix of .c and .cpp files. However, the
the -DINTEL_CUSTOMIZATION=1 flag is only being applied to files
compiled with 'icpx' and not for the files compiled with 'icx'.
This is because the compiler-rt is inheriting the flags from
the top level CMakeLists. This change adds the option to the
CMAKE_C_FLAGS as well, so that .c files compiled with 'icx' will
also get the definition.

[CMPLRLLVM-41722]