[HIRLocalityAnalysis] Keep track of assumed value for single blob in loop uppers for more consistent analysis (#8418)
Locality analysis uses a constant trip count for each loop in the loopnest
for analysis. This info can come from the loop itself but if there is no
info it assumed a trip count of 100.

In this PR, we also assign a constant value to the single blob in the upper
based on the assigned trip count. This lets us perform a more consistent
analysis when the blob is encountered during analysis as a blob coefficient
of IV. For example, we will assume %N = 5 and %M = 100 when computing
number of bytes accessed by A and B arrays in the loopnest below-

DO i1 = 0, %N-1, 1 <MAX_TC_EST = 5>
  DO i2 = 0, %M-1, 1 <assumed trip count of 100>
   A[%N * i2 + i1] = B[%M * i1 + i2]