Fix issues with ReorderFieldsOP (#8527)
[Opaque pointers] (CMPLRLLVM-41785) (CPU2017ref/523)

This pass shouldn’t check subtraction of a struct pointers are used in expected form when ReorderFieldOP is not triggered for the struct. Remove an assert that is checking for this.

Noticed that ReorderFields is triggered even when struct has no references. Use heuristics to avoid applying ReproderFields.

Fix DEBUG messages.