Update PGO unittests to pass with xmain (#8544)
* Update PGO unittests to pass with xmain

This updates the RUN lines for several PGO unittests because there
is a bug with the ld.gold in the binutils version that we used during
compiler builds that results in the following error when the
"--gc-sections" option is used:
  /usr/lib/gcc/x86_64-redhat-linux/8/../../../../lib64/crti.o(.gnu.build.attributes+0x14): error: relocation refers to local symbol "" [1], which is defined in a discarded section

This was a known bug https://sourceware.org/bugzilla/show_bug.cgi?id=23880
in ld.gold that has been fixed in later versions. On the r-drive, it
starts working properly with version 2.37. As this option is not critical
for functionality being tested in most cases, this PR removes that option
so that we can have working regression tests. On one test, the test
functionality required that option, so that test is being disabled
temporarily.

[CMPLRLLVM-41831]

* Add 'is'