[NFC] CMPLRLLVM-41169: Reorganize code for updating inline report calls/functions. (#8567)
* CMPLRLLVM-41169: Reorganize code for updating inline report calls/functions.

The code for updating the callsites and functions in the inlining report is
not completely clear in some places. This provides explicit get- interface
functions to get existing callsites and functions, add- interface functions
to add new callsites and functions, and getOrAdd- interface functions to add
a new callsite or function only if one does not already exist.

The updating is unified under makeCurrent(). Finally, initFunction now
creates and/or updates and existing function and all of its callsites.

* Clang format