[DTrans] Fix crash occurring when inline asm is seen by DTransFieldModRef (#8606)
* [DTrans] Fix crash occurring when inline asm is seen by DTransFieldModRef

A crash was occurring when inline asm is used within the IR because
this is implemented as a CallBase instruction, but the getCalledOperand()
function returns 'nullptr' for this case. This PR sets a conservative
result when inline asm is seen. It may be possible to relax this in the
future by checking the constraints to see if variables or memory are listed
in any of the constraints.

[CMPLRLLVM-41788]

* fix typo