[DTrans - OpaquePtr] Fix crash in PtrTypeAnalyzer seen when pointer to structure is indexed as an array type (#8664)
During one of the later passes of the PtrTypeAnalyzer, a crash was occuring in the
PtrTypeAnalyzer because many GEP instructions had been eliminated from the IR, and
the type that was indicated as being a pointer to a structure type was indexed
as an array type. This led to an assertion failure because indexing a structure
type requires constant values. This change will treat this IR as UNHANDLED by
DTRANS, as there is not currently a need to perform DTrans on any case that
has this issue.

[CMPLRLLVM-41896]