Align MultipleDevsCache with intel/llvm (#8806)
During previous pulldown we had a problem with this test and
cherry-picked several patches out of order. Seems there happened some
incorrect merge and test couldn't be build anymore. This patch just
takes version of this test from intel/llvm to align with open source and
to make it buildable again.