[AssumptionCache] Intel customization to findAffectedValues. (#8391)
[AssumptionCache] Intel customization to findAffectedValues.

This patch ensures that GlobalValues that are affected by assumptions
are properly added to the affected values map. Community LLVM does not
do this, so `assumptionsFor()` on any GlobalValue returns an empty list.

This patch is necessary for future work in VPlan where assumptions are
employed (i.e. 'omp simd aligned') -- without this patch, we cannot
obtain alignment assumptions for GlobalValues.