[VPlan][NFC] Native operand bundles for VPCallInstruction. (#8780)
[VPlan][NFC] Native operand bundles for VPCallInstruction.

This patch adds native operand bundles for VPCallInstruction. This means
that VP call instructions with attached operand bundles (from an
underlying call instruction) can be queried to obtain a
`VPOperandBundle` which corresponds to the range of `VPValue` operands
that correspond to the underlying bundle operands.

This patch is strictly additive, and preliminary: no functionality was
changed or removed. This is to enable future work (e.g.
VPAssumptionCache), and provide a path forward for more explicit
handling of operand bundles in the Vectorizer, if need be.

JR: CMPLRLLVM-42062