[Driver] Allow for IR compilation step for AMDGPU (#8896)
The Intel Compiler when built for releases restricts the ability to output
bitcode.  There is an additional step for -fsycl-targets=amdgcn that fails
due to this restriction.  Add -disable-intel-proprietary-opts to this step
to allow for the IR to be generated without error.

Cherry-pick of: https://github.com/intel-restricted/applications.compilers.llvm-project/pull/8891
Addresses:  CMPLRLLVM-42287