Add control over InstCombine GEP canonicalize swap transformation (#8882)
Performance was lost due on 538.imagick and 519.lbm due to transformations
not being triggered following the pulldown of the the community change
for https://reviews.llvm.org/D125845. This change caused GEP of GEPs
to be modified in a way that intermediate results of the GEP computation
no longer tracked to a particular structure field or array element, which
prevents analysis, such as DTrans, from tracking the accesses. This PR
inhibits that canocicalization in places where it was affecting DTrans,
when DTrans is enabled, in order to restore the benchmark performance.

[CMPLRLLVM-41970/CMPLRLLVM-42001]