[VPlan][TTI] declare simd parameter matching for multiple VFInfo objects (#8948)
* [VPlan][TTI] declare simd parameter matching for multiple VFInfo objects

This patch adds the ability to match declare simd functions based on multiple
VFInfo objects generated from the caller side. Previously, only one VFInfo was
created to match against potentially several "vector-variants". Previously,
since only one VFInfo was generated, we had to use a generic linear encoding
to try and match with the available "vector-variants". This was problematic
on the matching side in some cases because we wouldn't be able to determine
which variant to select. The other benefit is that this makes the matching
side explicit and easier to read. E.g., we match linearRef to linearRef,
linear to linear, etc.

* Address review comments