[VPlan] Disable SLP breakage heuristic for SIMD loops (#8949)
* [VPlan] Disable SLP breakage heuristic for SIMD loops

The SLP heuristic is motivated by the possibility that SLP vectorization
may provide better results than loop vectorization.  Therefore it should
only be used to decide whether or not loop vectorization occurs at all.
For #pragma vector always and #pragma omp simd loops, this decision has
already been made.  Disable the heuristic in such cases so that it can't
impact which VF is chosen.

Note that we still have an issue with the SLP heuristic causing us to
pick suboptimal VF for auto-vectorized loops.  This possibility is thus
far theoretical and won't be fixed in the short term.  The SLP heuristic
will reach end of life when we incorporate SLP into loop vectorization.

Reference:  CMPLRLLVM-42160