[DTrans - OpaquePtrs] Update more tests to opaque pointers (#8994)
* [DTrans - OpaquePtrs] Update more tests to opaque pointers

When the DTrans opaque pointer work began the compiler did not directly
support opaque pointers, and the LIT tests needed to be written as typed
pointers. Also, at the time, it was not clear whether the new opaque
pointer infrastructure was going to need to support operating on typed
pointers during the transition.

This PR continues migrating these LIT tests to use opaque pointers,
and updates the RUN lines that were there for typed pointers.
'bitcast ptr to ptr' instructions are left in place because the
infrastructure needs to be able to process these, as the compiler
can still generate them in some locations. For situations where
it was important to check the infrastructure when types were used
as competing types without bitcasts, there are already separate
tests written with opaque pointers.

[CMPLRLLVM-42235]

* Fix indentation on LIT test