Cherry-pick LIT test fixes to xmain-cand (#9011)
* Fix test fails in pulldown by adding -opaque-pointers

* 42373: Add -xmain-enable-gep0-removal option to match llorg. (#8987)

* 42375: Run update_test_checks on -O3 tests. (#8984)

* Add missing macro to interop-backend-traits.cpp (#7603)

This test generally uses traits of different backends when compiling
using the default spirv triple. It seems this means that the test
manually defines some macros that are usually defined by the backend.
For the ext_oneapi_cuda case a macro was missing leading to the test
failure. I have added this missing required macro.

Fixes this post-commit windows failure:
https://github.com/intel/llvm/actions/runs/3585225542/jobs/6032932018

Signed-off-by: JackAKirk <jack.kirk@codeplay.com>

Signed-off-by: JackAKirk <jack.kirk@codeplay.com>
Co-authored-by: Chang Lin <chang-sun.lin.jr@intel.com>
Co-authored-by: JackAKirk <jack.kirk@codeplay.com>