Added NoRecurse and default intrinsic attributes to subscript intrinsic (#8997)
I noticed that fortran functions are not deduced as norecurse due to presence of
subscript intrinsics which don't have the attribute. On investigating further I
realized that 'NoRecurse' property was not available for intrinsics and had to be created.

I found this old llvm thread on the topic-
https://groups.google.com/g/llvm-dev/c/BpI3jFGVXw4

Also decided to add the default attributes while investigating NoRecurse issue.