[DTrans - OpaquePtr] Fix assertion failure on DeleteFieldOP pass (#9111)
When the delete field opaque pointer pass updates global variable uses
within GEPOperators, it was expecting that the replacement would be a
new GEPOperator with just the indices changed. However, the
ConstantExpr::getGetElementPtr call does not return a GEPOperator in
the case when all the indices are zero, and instead just returns
the global variable. This would lead to an assertion failure, due to
the pass expecting to be able to treat the return value as a GEP.

[CMPLRLLVM-42545]