[HIRDeadStoreElimination] Replace store and substitutable loads with temp rather than propagating store RHS (#9000)
Now that we run constant/copy propagation after DSE, DSE analysis can focus
on legally substitutable loads rather than propagating store's RHS into the
load. We can let the propgation utility perform the cleanup.

This can help clean up more stores as it isn't always legal to propagate
store's RHS, for example, if it is a memref.

Theoretically, we may be losing out on some DSE opportunities because
propagating RHS on the fly can expose redundant node removal opportunities
(invoked by DSE on the fly) which in turn can result in more DSE. I
think the possibility of this happening is low and it can be mitigated
by running DSE again, if necessary.

This PR also improves the logic to recognize alloca as local to region by
ignoring 'store' uses of alloca.