[HIRNontemporalMarking] Add secondary store footprint threshold (#8709)
HIRNontemporalMarking uses a threshold based on the number of cache
lines used by the loop to determine which loops are large enough that
it makes sense to convert their stores to nontemporal stores. This can
be incomplete for some loops, so when using
-qopt-streaming-stores=always we disable this threshold to allow
streaming stores in more loops where the cache line count might be
difficult to calculate statically. However, disabling this threshold
also leads to peel/remainder loops generated by the vectorizer getting
converted to use nontemporal stores too.

To avoid that, this change adds a secondary threshold where stores are
not converted to nontemporal if their loop has a statically known
maximum trip count and the store size multiplied by that maximum trip
count is less than the threshold.

Jira: CMPLRLLVM-37776