Bugfix in usage of replaceStandAloneBlobByCanonExpr() (#9127)
This utility was currently asserting in some cases where we cannot perform
the replacement. The client did not have a way of knowing when it is safe
to call it so I introduced a canReplaceStandAloneBlobByCanonExpr() utility
which can be checked on the client side. This is now being used by copy
propagation utility.