CMPLRLLVM-42628: Fix Failing LIT tests due to redefintion of llvm.ptr.annotation (#9230)
The intrinsic llvm.ptr.annotation has been redefined to take two overloaded arguments.
This means that calls to getDefinition to create the definition and LIT tests that contain
the intrinsic must be written to accomodate the new form. For us, the second pointer
aryment can always be an i8* (typed version) or ptr (opaque version).