cmplrllvm-41501: missing set OMP_TGT_MAPTYPE_HOST_MEM type (#9229)
This is due to MapExpr is none.
The MapExpr is expression of used in the map clause, it could be null if
there is none.

When MapExpr is none, the check need to be the MapDecl same with
reduction VarDecl.  Basically, need to compare two VarDecls.

To do so, using
DenseMap<const ValueDecl *, const Expr *> ReductionExprs
instead:
DenseSet<const Expr *> ReductionExprs;

So that I can use VarDecl for comparsion.