[DTrans - OpaquePointers] Misc test case updates to use opaque pointers (#9267)
More conversion of typed pointers to opaque pointers in the LIT tests

'bitcast ptr to ptr' are kept in the analysis and safety tests
to ensure that bitcast instructions are still handled for type
identification, if they are encountered. These are removed from the
transform tests because they do not contribute to the results there.

A few tests are removed because they already had opaque pointer versions,
or are not relevant when typed pointers are not present.

A few updates were needed to the new-pm-lto-dtransop.ll test because
the AOS-to-SOA transformation only kicks in under certain conditions,
and removing the bitcast instruction from that test changed the
conditions.