Add serialization for Intel-added part of loop hint annotation token (#8748)
Fixes: https://jira.devtools.intel.com/browse/CMPLRLLVM-40421

* [clang] Add serialization for loop hint annotation tokens

When late parsed templates are used with PCH tokens are serialized. The
existing code does not handle annotation tokens which can occur due to
various pragmas.

This patch implements the serialization for annot_pragma_loop_hint.

This also enables use of OpenMP pragmas and #pragma unused which do not
need special serialization of the PtrData field.