[VPlan] Introduce native VPlanValueTracking. (#9194)
* [VPlan] Introduce native VPlanValueTracking.

This patch introduces native VPlanValueTracking: an analysis which is
computed from the VPlan IR itself, rather than as the result of calling
into the underlying LLVM ValueTracking (which may not always be
possible).

This works much in the same manner as LLVM's ValueTracking: bounded
recursion on the structure of IR values (i.e. recursing on instruction
operands), with constants and external values serving as base cases.
When possible, we utilize the underlying ValueTracking or
AssumptionCache to refine intermediate known bits.

This patch only adds preliminary support for basic operators, mostly
those for which LLVM already has convenient KnownBits interfaces for:
    - `add`, `sub`, `mul`, `udiv`
    - `and`, `or`, `xor`,
    - `lshr`, `ashr`, `shl`,
    - `srem`, `urem`
    - `trunc`, `sext`, `zext`

The intent is for this patch to lay the groundwork for subsequent
patches to fill in the rest of the gaps as-needed.

This replaces the previous implementation of
`VPlanValueTracking::computeKnownBits(const VPValue*, const VPInstruction*)`
which was stubbed out in the HIR path, and which simply deferred to
LLVM's ValueTracking in the LLVM-IR path.