[SYCL] Defer buffer release when no host memory to be updated (#9324)
SYCL2020 4.7.2.3. Buffer synchronization rules states that "A buffer can
be constructed from a range (and without a hostData pointer). The memory
management for this type of buffer is entirely handled by the SYCL
system. The destructor for this type of buffer does not need to block,
even if work on the buffer has not completed. Instead, the SYCL system
frees any storage required for the buffer asynchronously when it is no
longer in use in queues."
This commit implements this behavior for sycl::buffer.

This feature introduced more resources to be released in the end of
program if there was no chance to release them earlier. This commit
implements WA of known issues with global object destruction based on
thread_local usage, thread_local variables destroy earlier than global
variables that allow us to do release resources earlier.

Signed-off-by: Tikhomirova, Kseniya <kseniya.tikhomirova@intel.com>