[Driver] Cleanup Windows linking behaviors for Intel libs (#9410)
Changes from the community have caused a section of duplicate code when
dealing with adding the required default libraries for Windows linking.
This is causing a doubling up of the predefined macros (_MT, _DLL and _DEBUG)
as well as multiple inclusions of the MSVC runtime libs. The duplication of
the runtime libs is also causing a warning to be emitted from the linker due
to 2 different variants of the library being linked in when -fsycl
compilations are enabled.

Cleanup this portion of code, moving the SYCL specific additions to the new function.

Equivalent changes (without INTEL_CUSTOMIZED changes) are also being pushed
to intel/llvm.