Enclose input and scan phases of inscan loops in separate guard regions (#9321)
This patch changes creation scheme of guard directives (DIR.VPO.GUARD.MEM.MOTION)
for inscan reduction loops. Instead of wrapping the whole loop body,
input and scan phases of the loop are wrapped in guard region separately.

This aids in performing exclusive loop transformation which exchanges
input and scan loop phases between the loop. New scheme allows for
GEPs on reduction memory in the first phase not to be alive into the phase
that follows. This allows for easier handling of SSA structure when
performing the xform. Other values, except values on the separating pragma,
are not allowed to cross the separating pragma.

New guard region scheme is applicable only to inscan loops.
According to OMP specification, if a loop has inscan reduction-modifier used,
reduction clause without the inscan modifier must not appear on the same
construct. That means that we do not need to mix & match two guards
region creation approaches and they remain separated.

New inscan loops region scheme:

 LoopHeaderBB:
   <Loop header PHIs>
   %g1 = call @llvm.directive.region.entry() [ DIR.VPO.GUARD.MEM.MOTION() ]

   ...

   call @llvm.directive.region.exit(%g1) [ DIR.VPO.END.GUARD.MEM.MOTION() ]

   %s = call token @llvm.directive.region.entry() [ "DIR.OMP.SCAN"(), ... ]
   call void @llvm.directive.region.exit(token %s) [ "DIR.OMP.END.SCAN"() ]

   %g2 = call @llvm.directive.region.entry() [ DIR.VPO.GUARD.MEM.MOTION() ]

   ...

 LoopLatchBB::
   ...
   call @llvm.directive.region.exit(%g2) [ DIR.VPO.END.GUARD.MEM.MOTION() ]