[VPlan] VPAssumptionCache support for external assumptions. (#9239)
* [VPlan][AC] Support for importing LLVM assumptions external to VPlan
region.

Filter external assumptions in all paths so that external assumptions
(`AssumeInst*`) are only imported into the cache if they are valid. By
that we mean they:
    1. Are outside of the VPlan region,
    2. Dominate the VPlan region (loop entry).

This is done by templatizing `importExternalAssumptions()` on front-end
type, and adding a front-end interface `isValidExternalAssume` that
takes the assume and a dominator tree context (sufficent to call
`llvm::isValidAssumeForContext()`).