[Paropt][AtomicFreeRed] Use orig item's addrspace for replacement in teams rgn. (#9120)
* [Paropt][AtomicFreeRed] Use orig item's addrspace for replacement in teams rgn.

For non-SLM atomic-free reduction, when replacing the original value with
the privatized version -- i.e. an element of the teams reduction buffer,
the address space being used was "4" (GENERIC).

That is usually correct because the frontends always add a cast to
`addrspace 4` when emitting a clause. However, for globals, there is
a processing that happens in the prepare pass, which can strip off
constant-expr casts from `ptr addrspace (1) to ptr addrspace(4)`.
See canonicalizeGlobalVariableReferences() for reference.

After that processing, the clause operands for globals become
`addrspace(1)` instead of `addrspace(4)`, so when we do the replacement of
the original with the private copy, in some cases, we were replacing

  %0 = bitcast i32 addrspace(1)* %x to i8 addrspace(1)*

with:

  %0 = bitcast i32 addrspace(4)* %<private.copy> to i8 addrspace(1)*

causing illegal IR to be emitted.

We now use the original reduction item's addrspace for the replacement.

JIRA: CMPLRLLVM-42500