Enable LLVM IR output (#9507)
This patch enables LLVM IR output and removes several places where
proprietary optimizations were being disabled when IR output was
required. Also, the module flag to mark a module as having proprietary
optimizations which must restrict IR output is moved so that it applies
only to DTrans.

This unblocks CMPLRLLVM-38420.