[VPlan][MaskedModeLoops] Normalize masked mode loops bounds. (#9381)
* [VPlan][MaskedModeLoops] Normalize masked mode loops bounds.

The normalization of the loop bounds is needed to avoid the loop
index overflow in the mask check.
Currently we create the following loop for masked mode.

do i1 = LowerBound, UpperBound, VF {
   // Here i1 is a vector of {LowerBound + <0, ..., VF-1>}
   // and it can overflow if UpperBound is close to MAX, i.e. is
   // bigger than MAX - VF.
   if (i1 < UpperBound) {
      use_of(i1);
      ...
   }
   i1 += VF;
   allzero_check(i1 < UpperBound) // same here
}

After the change we generate the following:

 NewUpper = UpperBound - LowerBound;
 do i1 = 0, NewUpper, VF {
    adjusted_i1 = i1 + LowerBound;

    // Here i1 is a vector of {0 + <0, ..., VF-1>}
    // and it can't overflow as NewUpper is far from its MAX.
    if (i1 < NewUpper) {
       use_of(adjusted_i1);
       ...
    }
    i1 += VF;
    allzero_check(i1 < NewUpper) // same here
 }

This is not ideal solution, and it will not work in cases when
the whole loop with a big upper bound will be enforced to execute
in masked mode. But currently we create such loops only for peel
and remainder with a small number of iterations, so it can be
considered as a good temporary solution.