[VPlan][MinMax+index] Fix to account all linked values during importing. (#9475)
[VPlan][MinMax+index] Fix to account all linked values during importing.

When a reduction is under condition, the updating instructions are accounted
in reduction's linked value list and in the VPLoopEntityList they are not
linked to the reduction - we have only Exit and StartPhi linked there. This
causes problem with min/max+index reductions as they should be linked but
during importing the link is done through the update instructions. Adding
all update instructions to the VPLoopEntityList reduciton map fixes the
problem.