[VPlan] Disable SLP breakage heuristic in cost model (#9680)
Experiments have shown that the SLP breakage heuristic in the VPlan cost
model may no longer be making any difference.  Turning it off shows no
change in performance over a wide variety of benchmarks.  It seems likely
that the gather-scatter heuristic is covering much of the same ground.
This patch disables the heuristic but does not remove its code.  If there
is no fallout from QA testing, we can clean this up further.

Three tests are affected.  Two of these just test aspects of the SLP
breakage heuristic, so those are removed.  Without the SLP heuristic, one
loop in the invalid-alloca-store-use.ll test is unexpectedly vectorized,
so we disable vectorization for that loop to maintain the test's purpose.

Reference:  CMPLRLLVM-43181