review/dbudanov cmplr/CMPLRLLVM 43142 1 (#9768)
* [AAPointerInfo] track multiple constant offsets for each use

An expression of the form `gep(base, select(pred, const1, const2))` can result
in a set of offsets instead of just one. PointerInfo can now track these sets
instead of conservatively modeling them as Unknown. In general, AAPointerInfo
now uses AAPotentialConstantValues to examine the operands of the GEP.

Reviewed By: jdoerfert

Differential Revision: https://reviews.llvm.org/D138646

* Fix for CMPLRLLVM-43142.
https://reviews.llvm.org/D138646 introduces track multiple constant
offsets for each use for AAPointerInfo. A case when "offsets" are not
initialized is not covered in community code and causes assertion
fail later in AAPointerInfo::Access ctor. The fix add a handle for
this case and conservatively bails out attribute deduction.
Once the community started to handle this case the code in
this block can be removed.

---------

Co-authored-by: Sameer Sahasrabuddhe <sameer.sahasrabuddhe@amd.com>