[HIRScalarReplArray] Allow scalar replacement in the presence of adjacent refs (#9480)
In preparation for creating a loop independent scalar replacement pass out of
the current pass, I discovered a limitation where the current legality check
gives up in the presence of an 'adjacent' ref. This is a ref which either lies
outside the group due to distance threshold which would be common occurrence for
loop-independent scalar replacement or distance could not be computed due to
other subscripts in the ref.

Here's a simple example where scalar replacement will give up on the load
group {A[i], A[i+1]} due to adjacent store A[i+2] if distance threshold is set
to 1.

DO
 A[i+2] = A[i] + A[i+1];
END DO

More examples in lit tests.