[VPlan][VLS] Fix for conservative HIR based dependence checks that prevent VLS transformation
* [VPlan][VLS] Fix for conservative HIR based dependence checks that prevent VLS transformation

Consider the following HIR:

<10> t1 = a[2*i]
<20> t2 = b[...]
<30> t3 = a[2*i+1] + t2

Dependence checks to see if we can move <30> to <10> are done at HIR level.
This obviously cannot be done due to the dependence on t2. However, in VLS
transform we are not moving the whole HLInst but just the load of a[2*i+1].
It would be valid to do the following:

<10> t1 = a[2*i]
<11> tld = a[2*i+1]
<20> t2 = b[...]
<30> t3 = tld + t2

This is fixed by ignoring temp-to-temp dependences as a load instruction cannot
introduce a temp-to-temp dependency when trying to move it up.