[NFC][CMake] Run SYCL generators only as needed (#9794)
This is a cherry-pick of intel/llvm #7883, including some merge
conflict resolution due to different header paths.

The intent of this change is to reduce incremental build output and
time.

Use add_custom_target/command pairs to avoid generating build artifacts
repeatedly. The artifacts will still be regenerated if an input changes.