[VPlan][VLS] VLS improvements for performance parity with classic compiler. (#9766)
VLS transformation currently only handles groups where the group stride
matches the bit width of the group's access mask. As an example, a group
containing refs {a[2 * i], a[2 * i + 1], a[2 * i + 2]} is not handled.
ICC handles such a group by applying VLS optimization to refs {a[2 * i],
a[2 * i + 1] and doing a gather/scatter for a[2 * i + 2]. For now, do
the same in xmain by not considering a[2 * i + 2] for the group. This is
done by skipping over refs whose distance from any of the refs in the
group is >= group stride. This patch is part of changes needed to get
ifx/ifort performance parity for issue reported in CMPLRLLVM-39502.