CMPLRLLVM-33539: WRF/conus12km inlining extensions (#10076)
* CMPLRLLVM-33539: WRF/conus12km inlining extensions

We need to get some more cases inlined for WRF/conus12km. Single callsite functions which have the single dimensioned arrays passed to them also need to be inlined. Since the traversal of inlining callsites is done bottom up, we mark such callsites in the compile step and then inline them in the link step. This leads to more allocas in inlined functions, so I increased the default limit.

* Fix testing errors, add LIT test

* Fix bracketing

* More clang-format

* More clang-format

* Fix inlining report message

* Remove duplicate line