[DTrans] Fix crash in DTrans field mod/ref analysis when a literal struct type was a member (#10114)
* [DTrans] Fix crash in DTrans field mod/ref analysis when a literal struct type was a member

In the opaque pointer version of the DTrans field mod/ref analysis,
an assertion was being hit when marking fields as 'bottom' for
structures that didn't meet the safety conditions when the field
was a literal structure type. The marking of fields as bottom needs
to propagate the property to nested structures, however the lookup
of StructInfo objects is only supported for named structures, which
was leading to an assertion. This change marks all literal structure
fields as 'bottom' because we do not need to support these for the
case of interest, and avoids trying to lookup a StructInfo object
for a literal type later.

[CMPLRLLVM-43666]

* clang-format