Add vectorization passes in DPCPP O0 pipeline (#6696)
Patch allows vectorization passes to run in DPCPP with O0.
The idea was to give a customer an opportunity to disable sg-emulation and turn on a vectorization in DPCPP O0 instead.
In this part we enable Vectorization passes in DPCPP O0 pipeline and create a flag for it. The final result of the whole feature would be a replacement of SG emulation with vectorization (currently SG Emulation runs with O0 and we want vectorization run with O0 too). In the next patch we will enable vectorization and disable SG Emulation by default and it will require more specific tests (or editing old tests that were written for SG Emulation), work on other issues (if there will be some). But for now we just want to be sure that we do run required vec passes in the right order using a flag.

JIRA:CMPLRLLVM-31783