[OpenMP] Add InferAddressSpacesPass right before the vectorizer. (#9244)
Using -fopenmp-target-simd diverts the device code to IGC Vector
Backend which has a special lowering pass in xmain. That pass
is currently unable to deal with most of the generic addrspace
usecases (as is VC actually), so we help it by inferring addrspaces
before it. But as the inference pass is bad at processing vector IR
it's better to put it before VPlan vectorizer.

JIRA: CMPLRLLVM-33250