[VPlan] Generalize select-cmp reduction pattern for multi-AND conditions (#10175)
[VPlan] Generalize select-cmp reduction pattern for multi-AND conditions

The SelectICmp and SelectFCmp reduction handling recognizes either a
select or a phi as the exit instruction.  There is an existing assertion
that a phi, when present, has exactly two predecessors.  It is possible
for there to be three or more predecessors.  This patch removes the
assertion and generalizes the reduction handling for any number of
predecessors.

Reference: CMPLRLLVM-43459