CMPLRLLVM-43859: Exclude special dyn alloca callsites from limit. (#10184)
The previous change set caused regressions in the special inlining for WRF. We cannot increase the limit of normal dynamic allocas, as it will cause the shape95F/G_1_,main test case to overflow the stack at runtime. The best way to preserve the normal behavior while still getting the special cases is to exclude the special cases from the count. There aren't enough special cases to warrant putting a special limit there just for them.

* Add LIT tests