Fix annotation issue in inliner (#9896) (#10212)
JIRA: https://jira.devtools.intel.com/browse/CMPLRLLVM-42716

When Callee function is inlined into caller function, it is expected that any global annotation attached to the callee function be transferred to the caller function. There is currently no support in the inliner pass to do it. This change adds the logic to do this. All changes are INTEL customized.

Signed-off-by: Arvind Sudarsanam <arvind.sudarsanam@intel.com>