CMPLRLLVM-43093: Restrict SROA during paropt inlining for functions that will be inlined in loops (#9894)
This change set adds 2 new options for SROA pass: intel-preserve-cfg and
intel-modify-cfg. Both options are the same as preserve-cfg and modify-cfg,
but the enable an extra check that verifies if the function that will be
inlined into a loop. This will be used to run SROA during the inline
process of paropt. We only want to perform the optimization on functions
that will be inlined inside a loop to simplify the code.

This change set restores 6% regression for sycl_benchmarks_omp/aobench.