[Driver] ocloc split not properly taking splitting into account for OpenMP (#10150) (#10191)
When performing ocloc splitting for Windows, device code splitting is
enabled by default for SYCL based compilations.  This is not the case
for OpenMP compilations.  This is an opt-in type of behavior.  The
splitting behaviors were assuming the use of llvm-foreach, which caused
splitting with OpenMP to fail on Windows.

Update the behavior so the use of llvm-foreach is only implied based on
the known file types, and not just assuming it is enabled.

Addresses:  CMPLRLLVM-43834
Cherry-pick of:  https://github.com/intel-restricted/applications.compilers.llvm-project/commit/ecb23227bf53c8641ddd95f9574fd993ac903214