[VPlan][LoopVectorizationPlanner] Fix for CMPLRLLVM-43029 (#10172) (#10251)
Prevent trunc/zext -> and peephold optimization from overwriting
stride information where users of the and are subscripts. In DA,
improveStrideUsingIR() is called at the end of the initial propagation
of vector shapes, but is not called at the end of shape recomputation.
When the peephole optimization runs the shape of the new and is
propagated through to its users and overwrites the stride refinement
done in improveStrideUsingIR() and GS cost is assigned to loads/stores.
In this patch, we copy the stride of the zext instruction to the and
instead of calling the updateDivergence() interface to DA. That way,
the stride is retained and cost is seen as unit stride. Note: trying
to call improveStrideUsingIR() during shape recomputation won't work
because updateDivergence() will invalidate underlying IR and it will
bail out.