[OpaquePointers] Add a flag to disable opaque pointers only for SPIR-V/offload. (#10142)
The purpose of this flag is to allow the default of DPCPP_ENABLE_OPAQUE_POINTERS
to be flipped even before the offload code and other portions of the compiler
are fully ready.

With the exception of the change to libomptarget, this is essentially the same
change as intel/llvm#8073 absent merge conflicts. The
change to libomptarget is an implementation of CMPLRLLVM-40052 at the same time,
making that library respect the new flag.