[VecClone] Bail out when entry block contains array allocas (#10261)
VecClone splits the entry block to create a loop header, and ensures
the allocas are all moved to the entry block.  However, array allocas
can have arbitrary dependences on other instructions and memory, and
the present implementation doesn't handle that.  This patch implements
a bailout when array allocas are present.

Reference: CMPLRLLVM-43460