[ESIMD] CMPLRLLVM-43627 - Fix Coverity issues in ESIMD code (#10109) (#10201)
* [ESIMD] Fix usage of uninitialized field ESIMDVerifier::ForceStatelessMem (intel/llvm#8000)

The field was properly initialized in 'ESIMDVerifierPass` struct, but
not in `ESIMDVerifier`.
The first one gets the proper/guarantted value from
LangOpts.SYCLESIMDForceStatelessMem, while the second(i.e.
ESIMDVerifier) could not do the same as it is initialized from
llvm::createESIMDVerifierPass()

Fixing this issue showed that those SYCL methods that are allowed with
-fsycl-esimd-force-stateless-mem, can potentially and legally be used in
some other modes, for example, when used in LLVM IR consumed by 'opt'
component. Thus the meaning and name of the corresponding fields in
ESIMDVerifier and ESIMDVerifierPass structs were renamed from
ForceStatelessMem to MayNeedForceStatelessMemModeAPI.



* [ESIMD] Fix some stability and performance issues in code (intel/llvm#8049)

Performance issues are simple (copy overhead in for loops). Stability
issues are mostly usage of uninitialized variables. One of issues is
usage after std::move().
One if issues is usage of "Modified &= ..." where "Modified |= ..." was
supposed.

Signed-off-by: Vyacheslav N Klochkov <vyacheslav.n.klochkov@intel.com>