Enable opt-bisect in release builds (#10291)
Enable opt-bisect in release builds

Since we switched to the new pass manager, opt-bisect has been broken in
release builds. It only bisected codegen passes (which still use the
legacy pass manager). The code was simply excluded in release builds.

This patch reenables opt-bisect for release builds, except when DTrans
is enabled (in which case opt-bisect will be silently ignored).