[llvm-profdata] Report file version number on unsupported input (#10331)
* [llvm-profdata] Report file version number on unsupported input

The llvm-profdata data tool does not support backward compatibility
for raw profile data files. Instead, it just reports the cryptic
error message saying the file is an unsupported profile format
version. This is hard to diagnose, and there have been cases seen
where the llorg version of llvm-profdata is used instead of the
xmain tool because the xmain tool is not included as part of the
path because it conflicts with a system installed version. More
recently, a user encountered older files being generated, and
it was not clear what version the files were. This PR improves
the usability, by reporting the mismatched version numbers as
part of the error message.

[CMPLRLLVM-44284]

* clang-format

* Update comment