[VPlan][AssumptionCache] Add 'llvm.assume' operand to cache. (#10242)
Add the arg operand (i.e. `i1 %v` in `call void @llvm.assume(i1 %v)`) as
an affected value in the VPAssumptionCache. This additional information
that ValueTracking can use to compute known bits, e.g:

    %a = ...
    %cond = icmp eq %a, 500
    call void @llvm.assume(i1 %cond) <--- value of `%a` must be 500
                                          after this point

The above is not implemented in this patch; this work simply enables
such analysis to be implemented in the future.

JR: CMPLRLLVM-42918