[VPlan][DA] Simplify opaque pointer handling (#10465)
On entry to computeVectorShapeForSOAGepInst, we are guaranteed to have
two conditions that simplify things greatly. A VPSubscriptInst will
have at least one dimension, and any accesses through a VPSubscriptInst
will have the same size as the innermost element size. With these
guarantees in place, we can greatly simplify handling of opaque pointers.