[SLP] Preserve attributes for vectorized function calls. (#10124) (#10416)
Community vectorizers are not preserving attributes of the original call site
when handling calls. This can be problematic as different IMF attributes
implies different semantics. We fixed LoopVectorizer in f537518 but the
issue remains in SLP.

It gets more convoluted as in SLP multiple scalar calls are merged into one
single vector call, and the attributes of the original call sites can differ.
In this case we bail out vectorization of the call.

JIRA: CMPLRLLVM-43466