Update tests and NVPTX link behaviors (#10619)
A number of driver tests were failing due to disconnects between items
coming in from community and what is in intel/llvm or xmain.  Update tests
accordingly to match expectations, and fix problems with the linker used
for SYCL offloading with NVPTX to use fatbinary instead of the nvlink as
implied with changes from community.